Updated per RC review
