+++ src/widget/OktaSignIn.js
@@ -174,7 +174,8 @@ var OktaSignIn = (function () {
         The Okta Sign-In Widget is running in development mode.
         When you are ready to publish your app, embed the minified version to turn on production mode.
         See: https://developer.okta.com/code/javascript/okta_sign-in_widget#cdn
-      `);
+      `
+    );
 
     authClient = new OktaAuth({
       url: options.baseUrl,
