+++ webpack.common.config.js
@@ -1,8 +1,10 @@
+/* global module __dirname */
+
 var path      = require('path');
 var EMPTY     = 'widget/empty';
 var TARGET_JS = path.resolve(__dirname, 'target/js/');
 var SHARED_JS = TARGET_JS + '/shared';
-var plugins   = require('./webpack.plugins.config');
+var plugins   = require('./buildtools/webpack/plugins');
 
 // Return a function so that all consumers get a new copy of the config
 module.exports = function (outputFilename) {
@@ -137,9 +139,9 @@ module.exports = function (outputFilename) {
       setImmediate: false
     },
 
-    plugins: [
-      // Default to production mode
-      plugins.envPlugin('production')
-    ],
+    // Use the default plugins:
+    //   - Production mode (no console warnings)
+    //   - Not minified or uglified
+    plugins: plugins()
   };
 };
