+++ webpack.dev.config.js
@@ -1,8 +1,8 @@
 var common    = require('./webpack.common.config');
 var devConfig = common('okta-sign-in.js');
-var plugins   = require('./webpack.plugins.config');
+var plugins   = require('./buildtools/webpack/plugins');
 
 devConfig.devtool = '#inline-source-map';
-devConfig.plugins = [plugins.envPlugin('development')];
+devConfig.plugins = plugins({ debug: true });
 
 module.exports = devConfig;
